Added geometry tests
