+++ src/test/kotlin/UtilitiesTest.kt
@@ -1,6 +1,9 @@
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions.*
 
+internal class UtilitiesTest {
+}
+
 internal class HomMatrixTest{
 
     @Test
